+++ omc3/optics_measurements/measure_optics.py
@@ -96,7 +96,14 @@ def measure_optics(input_files: InputFiles, measure_input: DotDict) -> None:
             dispersion.calculate_normalised_dispersion(measure_input, input_files, beta_df, common_header)
 
     # Coupling ---
-    coupling.calculate_coupling(measure_input, input_files, phase_results, tune_dict, common_header)
+    try:
+        coupling.calculate_coupling(measure_input, input_files, phase_results, tune_dict, common_header)
+    except ZeroDivisionError:
+        if measure_input.accelerator.NAME in ('generic', 'sps'):
+            LOGGER.error("Coupling failed with zero-division error. Single plane BPMs? Skipped.")
+        else:
+            raise
+    
     if measure_input.only_coupling:
         return