+++ tests/unit/test_plot_spectrum.py
@@ -10,6 +10,7 @@ from omc3.plotting.plot_spectrum import main as plot_spectrum
 from omc3.plotting.spectrum.utils import PLANES, get_unique_filenames
 
 INPUT_DIR = Path(__file__).parent.parent / "inputs"
+INPUT_DIR_SPECTRUM_FILES = INPUT_DIR / "lhc_harpy_output"
 # Forcing non-interactive Agg backend so rendering is done similarly across platforms during tests
 matplotlib.use("Agg")
 
@@ -81,7 +82,7 @@ def test_no_tunes_in_files_plot(tmp_path, file_path, bpms):
 
     for f in glob(f"{file_path}*"):
         print(f)
-    for f in INPUT_DIR.glob(f"{file_path.name}*"):
+    for f in INPUT_DIR_SPECTRUM_FILES.glob(f"{file_path.name}*"):
         copy(f, tmp_path)
     file_path = tmp_path / file_path.name
     for p in PLANES:
@@ -91,6 +92,28 @@ def test_no_tunes_in_files_plot(tmp_path, file_path, bpms):
     plot_spectrum(
         files=[file_path], bpms=bpms, combine_by=["files", "bpms"],
     )
+
+
+@pytest.mark.basic
+def test_single_plane_bpms_stem_plot(tmp_path):
+    """ Use the SPS data to check that also single-plane BPMs can be plotted.
+    (this caused an error prior v0.21.0 as it was trying to plot `None`)
+    """
+    file_path = INPUT_DIR / "sps_data" / "lin_files" / "sps_200turns.sdds"
+    bpms = ["BPH.23608", "BPV.60108"]  # one from each plane
+    stem, waterfall = plot_spectrum(
+        files=[file_path],
+        output_dir=str(tmp_path),
+        bpms=bpms + ["unknown_bpm"],
+        lines_manual=[{"x": 0.44, "loc": "top", "text": "nothing here"}],
+        combine_by=["bpms"],
+    )
+    _, filename = list(get_unique_filenames([file_path]))[0]
+    filename = "_".join(filename)
+    assert len(list(_get_output_dir(tmp_path, file_path).iterdir())) == 1
+    assert len(waterfall) == 0
+    assert len(stem) == 1
+    assert (filename in stem) and isinstance(stem[filename], Figure)
 
 
 @pytest.mark.basic
@@ -115,7 +138,7 @@ def _get_output_dir(tmp_path, file_path):
 
 @pytest.fixture
 def file_path():
-    return INPUT_DIR / "spec_test.sdds"
+    return INPUT_DIR_SPECTRUM_FILES / "spec_test.sdds"
 
 
 @pytest.fixture