+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # OMC3 Changelog
 
+#### 2025-01-23 - v0.21.0 - _jdilly_, _fscarlier_, _fesoubel_
+
+- Fixed:
+  - Plot Optics Measurements: Added extra mpl style for clearer plots.
+  - LHC exciter BPM not found: Tells you which BPMs were searched for.
+  - Plot Spectrum: Correct error handling for Single-Plane BPMs.
+
+- Added:
+  - A `generic` accelerator class, that can be used for non-specifc accelerators.
+  - Global Correction: Total arc phase correction (`arc-by-arc`).
+  - Cleaning: Filter BPMs with NaNs.
+  - Cleaning: Log bad BPMs with reasons before raising errors.
+  - MAD-X wrapper: Inform about failed twiss.
+
 #### 2025-01-06 - v0.20.4 - _fsoubelet_
 
 - Fixed: